added params validation
